requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
Pillow>=10.0.0
piexif>=1.1.3
opencv-python>=4.8.0
anomalib>=0.3.0
torch>=2.0.0
torchvision>=0.15.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
langchain-openai>=0.0.10 